+++ changelog.md
@@ -8,7 +8,11 @@ compromise uses semver, and pushes to npm frequently
 
 While all _Major_ releases should be reviewed, our only two _large_ releases are **v6** in 2016 and and **v12** in 2019. Others have been mostly incremental, or niche.
 
-#### [Unreleased] 13.2.0
+#### [Unreleased]
+
+-move (some) contractions into tokenizer?
+
+#### 13.2.0
 
 - deprecate `.money()` and favour overloaded method in compromise-numbers plugin
 - add `.percentages()` and `.fractions()` to compromise-numbers plugin
@@ -22,6 +26,7 @@ While all _Major_ releases should be reviewed, our only two _large_ releases are
 - - remove conjugation and inflection functions
 - remove sourcemap from build process (too big)
 - improvements to `.numbers().units()`
+- fix for linked-list runtime error #744 with contractions
 
 #### 13.1.1